% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countneighbors.R
\name{count.neighbors}
\alias{count.neighbors}
\title{An evolClustR Function}
\usage{
count.neighbors(Cdata, radius)
}
\arguments{
\item{Cdata}{PDB file parsed to contain only central Carbon atoms}

\item{radius}{Radius size of cluster of interest}
}
\description{
Function to count number of neighbors for any given radius
}
\examples{
countneighbors()
}
\keyword{cluster}
