% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findneighbors.R
\name{find.neighbors}
\alias{find.neighbors}
\title{An evolClustR Function}
\usage{
find.neighbors(Cdata, radius, site)
}
\arguments{
\item{Cdata}{PDB file parsed to contain only central Carbon atoms}

\item{radius}{Radius of interest}

\item{site}{location of central site}
}
\description{
Function to return fractions in each cluster for any given radius
}
\examples{
find.neighbors()
}
\keyword{cluster}
