% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{fractions}
\alias{fractions}
\title{An evolClustR Function}
\usage{
fractions(Cdata, subs, radius)
}
\arguments{
\item{Cdata}{PDB file parsed to contain only central Carbon atoms}

\item{subs}{substited sites}

\item{radius}{Radius size of cluster of interest}
}
\description{
Function to return fractions in each cluster for any given radius
}
\examples{
fractions()
}
\keyword{cluster}
